refactor(ios): merge import orchestration into ImportPipeline
